requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
aniso8601==8.1.1
celery[redis]==5.0.5 
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
grpcio==1.35.0
grpcio-tools==1.35.0
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
itsdangerous==1.1.0
Jinja2==2.11.3
msgpack==1.0.2
MarkupSafe==1.1.1
numpy==1.22.3
#opencv-python==4.5.4.58
opencv-python-headless==4.2.0.34
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3 
requests==2.25.1
scikit-image==0.17.2
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1